new:usr:#3:Change package name. Set independent version to project
This commit changes the namespace to com.qubole.presto.udfs.
Also it sets the version to 0.1.0-SNAPSHOT. Setting the version required major
surgery of pom.xml file. I had to move dependencies from provided to a particular
version.
