+++ src/main/java/com/qubole/presto/udfs/WindowFunctionDefinition.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.qubole.udfs;
+package com.qubole.presto.udfs;
 
 import com.facebook.presto.spi.type.Type;
