+++ src/main/java/com/qubole/presto/udfs/window/BigintLastValueIgnoreNullFunction.java
@@ -11,29 +11,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.qubole.udfs.window;
+package com.qubole.presto.udfs.window;
 
 import com.facebook.presto.spi.type.Type;
-import com.facebook.presto.qubole.udfs.WindowFunctionDefinition;
+import com.qubole.presto.udfs.WindowFunctionDefinition;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
 
-import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
+import static com.facebook.presto.spi.type.BigintType.BIGINT;
 
 /**
  * Created by user on 13-03-2015.
  */
-public class BooleanFirstValueIgnoreNullFunction
+public class BigintLastValueIgnoreNullFunction
         extends FirstOrLastValueIgnoreNullFunction
         implements WindowFunctionDefinition
 {
-    public BooleanFirstValueIgnoreNullFunction(List<Integer> argumentChannels)
+    public BigintLastValueIgnoreNullFunction(List<Integer> argumentChannels)
     {
-        super(BOOLEAN, argumentChannels, Direction.FIRST);
+        super(BIGINT, argumentChannels, Direction.LAST);
     }
 
-    public BooleanFirstValueIgnoreNullFunction()
+    public BigintLastValueIgnoreNullFunction()
     {
         // Constructor is used by UdfFactory
         super();
@@ -42,18 +42,18 @@ public class BooleanFirstValueIgnoreNullFunction
     @Override
     public String getName()
     {
-        return "first_non_null_value";
+        return "last_non_null_value";
     }
 
     @Override
     public Type getReturnType()
     {
-        return BOOLEAN;
+        return BIGINT;
     }
 
     @Override
     public List<? extends Type> getArgumentTypes()
     {
-        return ImmutableList.<Type>of(BOOLEAN);
+        return ImmutableList.<Type>of(BIGINT);
     }
 }
