+++ src/main/java/com/qubole/presto/udfs/window/DoubleFirstValueIgnoreNullFunction.java
@@ -11,29 +11,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.qubole.udfs.window;
+package com.qubole.presto.udfs.window;
 
 import com.facebook.presto.spi.type.Type;
-import com.facebook.presto.qubole.udfs.WindowFunctionDefinition;
+import com.qubole.presto.udfs.WindowFunctionDefinition;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
 
-import static com.facebook.presto.spi.type.BigintType.BIGINT;
+import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
 
 /**
  * Created by user on 13-03-2015.
  */
-public class BigintLastValueIgnoreNullFunction
+public class DoubleFirstValueIgnoreNullFunction
         extends FirstOrLastValueIgnoreNullFunction
         implements WindowFunctionDefinition
 {
-    public BigintLastValueIgnoreNullFunction(List<Integer> argumentChannels)
+    public DoubleFirstValueIgnoreNullFunction(List<Integer> argumentChannels)
     {
-        super(BIGINT, argumentChannels, Direction.LAST);
+        super(DOUBLE, argumentChannels, Direction.FIRST);
     }
 
-    public BigintLastValueIgnoreNullFunction()
+    public DoubleFirstValueIgnoreNullFunction()
     {
         // Constructor is used by UdfFactory
         super();
@@ -42,18 +42,18 @@ public class BigintLastValueIgnoreNullFunction
     @Override
     public String getName()
     {
-        return "last_non_null_value";
+        return "first_non_null_value";
     }
 
     @Override
     public Type getReturnType()
     {
-        return BIGINT;
+        return DOUBLE;
     }
 
     @Override
     public List<? extends Type> getArgumentTypes()
     {
-        return ImmutableList.<Type>of(BIGINT);
+        return ImmutableList.<Type>of(DOUBLE);
     }
 }
