+++ src/main/java/com/qubole/presto/udfs/window/VarcharLastValueIgnoreNullFunction.java
@@ -11,29 +11,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.qubole.udfs.window;
+package com.qubole.presto.udfs.window;
 
-import com.facebook.presto.qubole.udfs.WindowFunctionDefinition;
 import com.facebook.presto.spi.type.Type;
+import com.qubole.presto.udfs.WindowFunctionDefinition;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
 
-import static com.facebook.presto.spi.type.BigintType.BIGINT;
+import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
 
 /**
  * Created by user on 13-03-2015.
  */
-public class BigintFirstValueIgnoreNullFunction
+public class VarcharLastValueIgnoreNullFunction
         extends FirstOrLastValueIgnoreNullFunction
         implements WindowFunctionDefinition
 {
-    public BigintFirstValueIgnoreNullFunction(List<Integer> argumentChannels)
+    public VarcharLastValueIgnoreNullFunction(List<Integer> argumentChannels)
     {
-        super(BIGINT, argumentChannels, Direction.FIRST);
+        super(VARCHAR, argumentChannels, Direction.LAST);
     }
 
-    public BigintFirstValueIgnoreNullFunction()
+    public VarcharLastValueIgnoreNullFunction()
     {
         // Constructor is used by UdfFactory
         super();
@@ -42,18 +42,18 @@ public class BigintFirstValueIgnoreNullFunction
     @Override
     public String getName()
     {
-        return "first_non_null_value";
+        return "last_non_null_value";
     }
 
     @Override
     public Type getReturnType()
     {
-        return BIGINT;
+        return VARCHAR;
     }
 
     @Override
     public List<? extends Type> getArgumentTypes()
     {
-        return ImmutableList.<Type>of(BIGINT);
+        return ImmutableList.<Type>of(VARCHAR);
     }
 }
