+++ src/main/resources/META-INF/services/com.facebook.presto.spi.Plugin
@@ -1,1 +1,1 @@
-com.facebook.presto.qubole.udfs.UdfPlugin
+com.qubole.presto.udfs.UdfPlugin
